% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{bivariate_plot}
\alias{bivariate_plot}
\title{Plot the measured concentration (or dilution) vs a variable defined by the user.}
\usage{
bivariate_plot(Ds, YResp = "MeasuredConcentration",
  YLongName = "Measured response")
}
\arguments{
\item{Ds}{is a data.frame with 6 columns (RunTechnician, ConcentrationLabel, ConcentrationValue,
CalibCurve, ReplicateNumber, Responses). Or a matrix with 7 columns (RunTechnician, ConcentrationLabel, ConcentrationValue,
ReplicateNumber, Status, Responses, MeasuredConcentration).
Responses column might have been filled by simulation or with the experimental results.}

\item{YResp}{a string of character corresponding to the name of the y axis.}

\item{YLongName}{a string of character corresponding to the label of the y axis (the measured concentration).
 @examples
 monDOE = DOE_Run_Repl_Conc(nRun = 3, nreplicates = 2, ConcVect = c(0, 50, 100, 125,150, 175, 200 ), Threshold='NaN', factorlist=c('RunTechnician','ConcentrationLabel','ConcentrationValue','ReplicateNumber','Status','Response'))
 TPL <- norm_lin_dataset_sim(monDOE, intercept = 1, slope = 2, SDrun=1.5, SDrep=3, biais = 0)
bivariate_plot(TPL)

monDOE = DOE_Run_Repl_Conc(nRun = 7, nreplicates = 2, ConcVect = c(0, 50, 100, 125,150, 175, 200 ), Threshold='NaN', factorlist=c('RunTechnician','ConcentrationLabel','ConcentrationValue','ReplicateNumber','Status','Response'))
TPL <- norm_lin_dataset_sim(monDOE, intercept = 1, slope = 2, SDrun=1.5, SDrep=3, biais = 0)
bivariate_plot(TPL)}
}
\description{
Plot the measured concentration (or dilution) vs a variable defined by the user.
}
