% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calib_diagnostics}
\alias{calib_diagnostics}
\title{Provides and displays diagnostics}
\usage{
calib_diagnostics(PredResid, LOFpValVect)
}
\arguments{
\item{PredResid}{is a matrix with two columns corresponding to the fitted values and the residuals.}

\item{LOFpValVect}{is a numeric vector with the p-values of the lack of fit test for each calibration curves.}
}
\value{
N/A text and plot for Rmarkdown
}
\description{
calib_diagnostics returns diagnostics to assess the validity of the calibration model.
}
\examples{
data(calib_data)
RespCalib = calib_coef(DS = calib_data,LowerRange=0,UpperRange=200,MyModel='lm')
calib_diagnostics(RespCalib$PredResid, RespCalib$LOFpValVect)
}
