% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calib_plot}
\alias{calib_plot}
\title{Displays the calibration curves}
\usage{
calib_plot(Ds)
}
\arguments{
\item{Ds}{is a data.frame with 6 columns (RunTechnician, ConcentrationLabel, ConcentrationValue, CalibCurve, ReplicateNumber,
Responses). Responses column might have been filled by simulation or with the experimental results.}
}
\value{
N/A plots for Rmarkdown or shiny
}
\description{
calib_plot plots the calibration curves.
}
\examples{
data(calib_data)
calib_plot(calib_data)

myDOE =calib_doe(nRun = 2, nCalibCurvesPerRun = 2, nrepCalib = 2,
ConcVect = c(0, 50, 100, 125, 150, 175, 200))
calib_data2 = norm_lin_dataset_sim(myDOE, intercept = 0, slope = 1, SDrun = 0.3,
SDrep = 0.3, biais = 0)
calib_plot(calib_data2)

myDOE =calib_doe(nRun = 60, nCalibCurvesPerRun = 60, nrepCalib = 3,
ConcVect = c(0, 50, 100, 125, 150, 175, 200))
calib_data3 = norm_lin_dataset_sim(myDOE, intercept = 0, slope = 1, SDrun = 0.3,
SDrep = 0.3, biais = 0)
calib_plot(calib_data3)

}
