% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{display_a_DOE}
\alias{display_a_DOE}
\title{Display the layout of the design of experiment of a dataset in a contengency table.}
\usage{
display_a_DOE(Ds1, GroupVar = "CalibCurve",
  GroupVar2 = "ConcentrationValue")
}
\arguments{
\item{Ds1}{a matrix or a data.frame produced by a DOE generating function or by a generation of simulated observations function.}

\item{GroupVar}{is a string of character corresponding to a name of categorical variable (i.e. a column) of Ds1.
Each level of this variable will be a row of the contengency table.}

\item{GroupVar2}{is a string of character corresponding to a name of categorical variable (i.e. a column) of Ds1.
Each level of this variable will be a column of the contengency table.}
}
\value{
An output of class 'table' which corresponds to the contengency table of the two selected variables.
}
\description{
Display the layout of the design of experiment of a dataset in a contengency table.
}
\examples{
data(ref_calib_doe)
display_a_DOE(ref_calib_doe, GroupVar = 'CalibCurve', GroupVar2 = 'ConcentrationValue')
}
