% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOE.R
\name{DOE_Run_Repl_Conc}
\alias{DOE_Run_Repl_Conc}
\title{Generate the design of experiment for validation studies}
\usage{
DOE_Run_Repl_Conc(nRun = 3, nreplicates = 2, ConcVect = c(0, 50, 100, 125,
  150, 175, 200), Threshold = "NaN", factorlist = c("RunTechnician",
  "ConcentrationLabel", "ConcentrationValue", "ReplicateNumber", "Status",
  "Response"))
}
\arguments{
\item{nRun}{is a numirical scalar corresponding to the number of run(s) sessions
that is (are) orginised in the design of experiment.}

\item{nreplicates}{is a numirical scalar corresponding to the number of replicate
in the design of experiment.}

\item{ConcVect}{is a vector of numerical values corresponding to the concentration
(or dilution) in the design of experiment.}

\item{Threshold}{is a numerical scalar corresponding to the detection limit or
the decision limit.}

\item{factorlist}{is a vector of characters corresponding to the name of the columns
of the DOE Do not change!}
}
\value{
An output of class 'matrix' which corresponds to the full factorial design
of experiment that is run(s) x replicate(s) x concentration level(s).
}
\description{
Generate the design of experiment for validation studies
}
\examples{
DOE_Run_Repl_Conc(nRun = 3, nreplicates = 2, ConcVect = c(0, 50, 100, 125,150, 175, 200 ),
 Threshold='NaN',
 factorlist=c('RunTechnician','ConcentrationLabel',
'ConcentrationValue','ReplicateNumber','Status','Response'))
}
