% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{find_column}
\alias{find_column}
\title{Retrieve the data of a column form a matrix or a data.frame.}
\usage{
find_column(Ds0, NomdeColonne, Output = "Vector")
}
\arguments{
\item{Ds0}{matrix or a data.frame}

\item{NomdeColonne}{the name of the column of interest}

\item{Output}{(optional) export the data of the column as a vector or only the number of the column.
@examples
ExMatrix = cbind(c(11,12,13),c(21,22,23),c(31,32,33))
colnames(ExMatrix) <- c('first_col','second_col','third_col')
find_column(ExMatrix, NomdeColonne='second_col', Output = "Vector")
find_column(ExMatrix, NomdeColonne='second_col', Output = "Number")}
}
\description{
Retrieve the data of a column form a matrix or a data.frame.
}
