% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_link.R
\name{linear_link}
\alias{linear_link}
\title{Linear link function gives the response (i.e. y or ordinate) of a linear model.}
\usage{
linear_link(intercept = 0, slope = 1, x)
}
\arguments{
\item{intercept}{is a real number corresponding to the intercept of the linear model.}

\item{slope}{is a real number corresponding to the slope of the linear model.}

\item{x}{is a real number (or a vector of real numbers) corresponding the abscissa (the input) of the linear model.}
}
\value{
y, the output is an real number (or a vector of real numbers) corresponding to the ordinate(s) of the linear model.
}
\description{
Linear link function gives the response (i.e. y or ordinate) of a linear model.
}
\examples{
linear_link(intercept = 0, slope=1,x=c(1:3))
}
