% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_link.R
\name{linear_link_inv}
\alias{linear_link_inv}
\title{Inverse linear link function gives the abscissa (x) of a linear model.}
\usage{
linear_link_inv(intercept = 0, slope = 1, y)
}
\arguments{
\item{intercept}{is a real number corresponding to the intercept of the linear model.}

\item{slope}{is a real number corresponding to the slope of the linear model.}

\item{y}{is a real number (or a vector of real numbers) corresponding the ordinate(s) of the linear model.}
}
\value{
x, the output is an real number (or a vector of real numbers) corresponding to the abscissa of the linear model.
}
\description{
Inverse linear link function gives the abscissa (x) of a linear model.
}
\examples{
linear_link(intercept = 0, slope=1,x=c(1:3))
}
