% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{norm_lin_dataset_sim}
\alias{norm_lin_dataset_sim}
\title{Generate simulated observations following a linear model with normal random error.}
\usage{
norm_lin_dataset_sim(MyDOE, intercept = 0, slope = 1, SDrun = 10,
  SDrep = 10, biais = 0)
}
\arguments{
\item{MyDOE}{is a data.frame with 6 columns (RunTechnician, ConcentrationLabel,+
ConcentrationValue, CalibCurve, ReplicateNumber,
Responses). This data.frame is produced by a DOE generating function.}

\item{intercept}{the intercept of the linear model.}

\item{slope}{the slope of the linear model.}

\item{SDrun}{the standard deviation between analytical runs.}

\item{SDrep}{the standard deviation between replicates (within an analytical run).}

\item{biais}{the bias between the true values and the simulated data.}
}
\value{
The output is a data.frame with 6 columns (RunTechnician, ConcentrationLabel,
 ConcentrationValue, CalibCurve, ReplicateNumber,Responses).
Responses column is filled with the simulated data following a linear model with normal random error.
}
\description{
Generate simulated observations following a linear model with normal random error.
}
\examples{
MyDOE <- calib_doe(nRun = 2, nCalibCurvesPerRun = 3, nrepCalib = 5,
  ConcVect = c(0,50, 100, 125, 150, 175, 200))

CalibDs <- norm_lin_dataset_sim(MyDOE, intercept = 1, slope = 2, SDrun=1.5,
 SDrep=3, biais = 0)
}
\seealso{
DOE generating functions such as \code{\link{calib_doe}}
}
