% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_sim_tot_var}
\alias{ref_sim_tot_var}
\title{Example of an empty calibration dataset.}
\format{A data frame with 210 rows and 6 variables:
\describe{
  \item{RunTechnician}{A string of character with the name of the run (lab session)}
  \item{ConcentrationLabel}{A string of character with the lable of the concentration,
   eg. 'Blanc' or '050'}
  \item{ConcentrationValue}{A numeric corresponding to the concentration value}
  \item{CalibCurve}{A string of character corresponding to the id of the calibration curve}
  \item{ReplicateNumber}{A string of character corresponding to the id of the replicate}
  \item{Response}{A numeric value corresponding to the observed measurement, in equipment unit.
  If there is no observation the response as a NaN value.}
}}
\usage{
ref_sim_tot_var
}
\description{
A dataset containing an design of experiment for calibration assessment (with empty results).
Summary: 1000 runs, nCalibCurvesPerRun = 1 nrepCalib = 1, Only 1 concentration 10 ppm.
}
\details{
ref_sim_tot_var
}
\keyword{datasets}
