% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{summary_stat}
\alias{summary_stat}
\title{Summarize the dataset through descriptive statistics}
\usage{
summary_stat(Ds0, VarVect, GroupVar = "")
}
\arguments{
\item{Ds0}{matrix or a data.frame}

\item{VarVect}{a vector with the name(s) of the numerical variable(s) to summerized}

\item{GroupVar}{(optional) a string of character corresonding a categorical variable that can
be used to apply the summary statistics to a subgroups of observations
of the dataset.
@examples
data(calib_data)
summary_stat(Ds0=calib_data,VarVect=c('MeasuredConcentration'), GroupVar= 'ConcentrationLabel')}
}
\description{
Summarize the dataset through descriptive statistics
}
